updates per PR feedback
* update 'porter bundle list' verbiage
* relocate list implementation into porter (remove from cnab/provider)
* print times in a condensed, human-readable format
