+++ pkg/porter/cnab.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/deislabs/porter/pkg/config"
 	"github.com/deislabs/porter/pkg/parameters"
-	"github.com/deislabs/porter/pkg/printer"
 	"github.com/pkg/errors"
 
 	cnabprovider "github.com/deislabs/porter/pkg/cnab/provider"
@@ -19,7 +18,6 @@ type CNABProvider interface {
 	Install(arguments cnabprovider.InstallArguments) error
 	Upgrade(arguments cnabprovider.UpgradeArguments) error
 	Uninstall(arguments cnabprovider.UninstallArguments) error
-	List(opts printer.PrintOptions) error
 }
 
 // sharedOptions are common options that apply to multiple CNAB actions.
